+++ client/src/components/layouts/index.ts
@@ -1,0 +1,2 @@
+export * from "./HomeLayout";
+export * from "./Layout";
